# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Geto API",
    "description": "Geto Database Management Tool API",
    "license": {
      "name": ""
    },
    "version": "0.4.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "crate::auth::handlers"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatus"
                }
              }
            }
          },
          "401": {
            "description": "Invalid password"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthStatus": {
        "type": "object",
        "required": [
          "authenticated"
        ],
        "properties": {
          "authenticated": {
            "type": "boolean"
          }
        }
      },
...
      "LoginInput": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
...
    }
  }
}